refactor(shaderpark): extract comment completion suppression helper
